==> Scripts/s11_Resting_state_FC_feature_selection_and_filtering.ipynb <==
import os
import pandas as pd
import numpy as np
import glob
from scipy.stats import pearsonr

#path to folders
path_m = '/media/DataD800/Alina/' #main set
path_r = '/media/DataD800/Alina/retest_set/' #retest set

# univariate feature filtering approach like in Dubois 2018, skip feature if  p > 0.01

#load tables with cognition and restFC

df_iq = pd.read_csv(path_m+'MLtables/cognition_table.csv', index_col=0)['CogTotalComp_Unadj']
df_iq = df_iq.dropna()
df_rest = pd.read_csv(path_m+'MLtables/rest_table.csv', index_col=0)

#excludeing everything with p-value bigger than 0.01
cn=[]
val = []
for col in df_rest.columns:
    p = pearsonr(df_rest.loc[df_iq.index, col], df_iq)[1]
    val+=[p]
    if p < 0.01:
        cn+=[col]

print(len(cn))

#filter to chosen columns
df_rest_filt = df_rest.loc[:,cn]
#and save
df_rest_filt.to_csv(path_m+'MLtables/rest_table_featfiltered.csv')

df_rest_filt.shape

#filtering a retest set according to chosen ROIs

df_rest_retest = pd.read_csv(path_r+'MLtables/rest_table.csv', index_col=0)
df_rest_filt_retest = df_rest_retest.loc[:,cn]
df_rest_filt_retest.to_csv(path_r+'MLtables/rest_table_featfiltered.csv')
